+++ app/scripts/utils/color-to-hex.js
@@ -1,15 +1,21 @@
 import { color } from 'd3-color';
-import * as PIXI from 'pixi.js';
+
+// Configs
+import GLOBALS from '../configs/globals';
 
 /**
  * Convert a regular color value (e.g. 'red', '#FF0000', 'rgb(255,0,0)') to a
  * hex value which is legible by PIXI
  */
 const colorToHex = colorValue => {
-  if (!colorValue) return PIXI.utils.rgb2hex([0, 0, 0]);
+  if (!colorValue) return GLOBALS.PIXI.utils.rgb2hex([0, 0, 0]);
 
   const c = color(colorValue);
-  const hex = PIXI.utils.rgb2hex([c.r / 255.0, c.g / 255.0, c.b / 255.0]);
+  const hex = GLOBALS.PIXI.utils.rgb2hex([
+    c.r / 255.0,
+    c.g / 255.0,
+    c.b / 255.0
+  ]);
 
   return hex;
 };